// src/commons/enums.ts
export enum ImageType {
    PRIMARY = 'PRIMARY',
    SECONDARY = 'SECONDARY',
}

export enum PromotionType {
    MEMBER = 'MEMBER',
    GUEST = 'GUEST',
}

export enum CancellationType {
    FREE_CANCELLATION = 'FREE_CANCELLATION',
    NOT_REFUNDABLE = 'NOT_REFUNDABLE',
}

export enum RatingType {
    SELF = 'SELF',
    STAR = 'STAR',
}

export enum SortingOption {
    HIGH_LOW = 'Price high-low',
    LOW_HIGH = 'Price low-high',
}